+++ lib/src/confiot/confiot.dart
@@ -1,0 +1,12 @@
+library confiot;
+
+import 'package:collection/collection.dart';
+import 'package:flutter/foundation.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:layrz_models/layrz_models.dart';
+
+part 'confiot.freezed.dart';
+part 'confiot.g.dart';
+
+part 'src/file.dart';
+part 'src/namespaces.dart';